Separate benchmarks in multiple files
